+++ packages/browser/lib/ui/card.ts
@@ -58,6 +58,8 @@ export default class Card {
         hiddenFields: (this.#options.hiddenFields ?? [])?.join(","),
         fields: this.#options.fields,
         acceptedBrands: this.#options.acceptedBrands,
+        autoComplete: this.#options.autoComplete,
+        autoProgress: this.#options.autoProgress,
       },
     };
   }