+++ packages/ui-components/src/Card/index.tsx
@@ -16,7 +16,12 @@ import { CardHolder } from "./CardHolder";
 import { CardNumber } from "./CardNumber";
 import { DEFAULT_TRANSLATIONS } from "./translations";
 import { useCardReader } from "./useCardReader";
-import { changePayload, isAcceptedBrand, swipePayload } from "./utilities";
+import {
+  autoProgress,
+  changePayload,
+  isAcceptedBrand,
+  swipePayload,
+} from "./utilities";
 import type { CardForm, CardConfig } from "./types";
 import type { CardFrameClientMessages, CardFrameHostMessages } from "types";
 
@@ -95,6 +100,10 @@ export function Card({ config }: { config: CardConfig }) {
       },
     },
     onChange: (formState) => {
+      if (config?.autoProgress) {
+        autoProgress(formState);
+      }
+
       const triggerChange = async () => {
         if (!ev) return;
         const cardData = await changePayload(ev, formState, fields);
@@ -161,6 +170,7 @@ export function Card({ config }: { config: CardConfig }) {
             autoFocus={config.autoFocus}
             placeholder={t("name.placeholder")}
             value={form.values.name}
+            autoComplete={config.autoComplete?.name ?? true}
             {...form.register("name")}
           />
           {form.errors?.name && (
@@ -184,6 +194,7 @@ export function Card({ config }: { config: CardConfig }) {
             autoFocus={config.autoFocus}
             placeholder={t("number.placeholder")}
             value={form.values.number}
+            autoComplete={config.autoComplete?.number ?? true}
             {...form.register("number")}
           />
           {form.errors?.number && (
@@ -206,6 +217,7 @@ export function Card({ config }: { config: CardConfig }) {
             disabled={!config}
             readOnly={cardReaderListening}
             placeholder={t("expiry.placeholder")}
+            autoComplete={config.autoComplete?.expiry ?? true}
             {...form.register("expiry")}
           />
           {form.errors?.expiry && (