+++ e2e-apps/browser/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @evervault/browser-e2e-app
 
+## 1.0.3
+
+### Patch Changes
+
+- Updated dependencies [3156568]
+  - @evervault/browser@2.28.0
+
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies [1df9ba2]
+- Updated dependencies [e2f0721]
+  - @evervault/browser@2.27.0
+
 ## 1.0.1
 
 ### Patch Changes