+++ examples/next-3ds/CHANGELOG.md
@@ -1,5 +1,19 @@
 # example-next-3ds
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies [3156568]
+  - @evervault/react@2.9.0
+
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies [1df9ba2]
+  - @evervault/react@2.8.0
+
 ## 0.1.1
 
 ### Patch Changes