+++ packages/types/CHANGELOG.md
@@ -1,5 +1,18 @@
 # types
 
+## 0.3.0
+
+### Minor Changes
+
+- 3156568: Add autoProgress option to Card component to automatically progress to the next input when an input becomes valid.
+
+## 0.2.0
+
+### Minor Changes
+
+- 1df9ba2: Add functionality to disable autoCompletion for fields in the Card component
+- e2f0721: Adds 3DS method support
+
 ## 0.1.0
 
 ### Minor Changes