+++ packages/ui-components/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @evervault/ui-components
 
+## 1.10.0
+
+### Minor Changes
+
+- 3156568: Add autoProgress option to Card component to automatically progress to the next input when an input becomes valid.
+
+### Patch Changes
+
+- Updated dependencies [3156568]
+  - @evervault/react@2.9.0
+  - types@0.3.0
+  - @evervault/card-validator@1.0.5
+  - shared@1.0.6
+
+## 1.9.0
+
+### Minor Changes
+
+- 1df9ba2: Add functionality to disable autoCompletion for fields in the Card component
+- e2f0721: Adds 3DS method support
+
+### Patch Changes
+
+- Updated dependencies [1df9ba2]
+- Updated dependencies [e2f0721]
+  - @evervault/react@2.8.0
+  - types@0.2.0
+  - @evervault/card-validator@1.0.5
+  - shared@1.0.5
+
+## 1.8.1
+
+### Patch Changes
+
+- a23d2f5: Add content type header to request to fetch 3DS session
+
 ## 1.8.0
 
 ### Minor Changes